requirements.txt
django 
djangorestframework
requests
notebook
pandas
python-dotenv
datasets
PILLOW
transformers
scikit-learn
jupyter
huggingface_hub
numpy
django-cors-headers
scipy